setup the files and move things around
