+++ social/network/api/src/commonMain/kotlin/xyz/ksharma/krail/social/network/api/SocialManager.kt
@@ -1,0 +1,5 @@
+package xyz.ksharma.krail.social.network.api
+
+interface SocialManager {
+
+}
