+++ social/network/api/src/commonMain/kotlin/xyz/ksharma/krail/social/network/api/model/SocialType.kt
@@ -1,0 +1,8 @@
+package xyz.ksharma.krail.social.network.api.model
+
+enum class SocialType(val displayName: String) {
+    LinkedIn(displayName = "linkedin"),
+    Reddit(displayName = "reddit"),
+    Instagram(displayName = "instagram"),
+    Facebook(displayName = "facebook"),
+}
